+++ src/tableView.js
@@ -40,15 +40,21 @@ Handsontable.TableView = function (instance) {
   });
 
   $(document.documentElement).on('mousedown', function (event) {
-    if (that.instance.getSettings().outsideClickDeselects) {
-      var next = event.target;
+    var next = event.target;
+    if (next !== that.wt.wtTable.spreader) { //immediate click on "spreader" means click on the right side of vertical scrollbar
       while (next !== null && next !== document.documentElement) {
-        if (next === instance.rootElement[0] || $(next).attr('id') === 'context-menu-layer' || $(next).is('.typeahead li')) {
+        if (next === instance.rootElement[0] || $(next).attr('id') === 'context-menu-layer' || $(next).is('.context-menu-list') || $(next).is('.typeahead li')) {
           return; //click inside container
         }
         next = next.parentNode;
       }
+    }
+
+    if (that.instance.getSettings().outsideClickDeselects) {
       that.instance.deselectCell();
+    }
+    else {
+      that.instance.destroyEditor();
     }
   });
 
@@ -207,6 +213,12 @@ Handsontable.TableView = function (instance) {
     that.wt.update('height', that.containerHeight);
     that.instance.forceFullRender = true;
     that.render();
+  });
+
+  $(that.wt.wtTable.spreader).on('mousedown.handsontable, contextmenu.handsontable', function (event) {
+    if(event.target === that.wt.wtTable.spreader && event.which === 3) { //right mouse button exactly on spreader means right clickon the right hand side of vertical scrollbar
+      event.stopPropagation();
+    }
   });
 };