+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Bugfixes:
 - scrolling issues (#343, #331, partially #360)
 - deselection when param `outsideClickDeselects` set to true (issue #351)
 - order of callbacks when clicked outside of a cell (issue #269)
+- context menu was shown when clicked outside of the table (issue #306)
 
 ## [0.8.2](https://github.com/warpech/jquery-handsontable/tree/v0.8.2) (Jan 7, 2013)